pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "limen"
version = "0.1.0"
description = "Framework de raisonnement adaptatif par seuil de compétence pour systèmes IA"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Open Research Starter" }
]
dependencies = [
  "pydantic>=2.8.0",
  "PyYAML>=6.0.1",
  "requests>=2.32.0",
  "python-dotenv>=1.0.1",
  "openai>=1.40.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["limen*"]
